@@ 4.8 Discovering deeper abstractions (p. 96)
End: Just before ## 4.9 (p. 102)

Using Flocking Rules we've unearthed a deeper abstraction.

The first *clue* that we need to create another method (that the obvious
*abstraction is wrong*), was that we'd be
implementing three conditional branches, vs. 2 that we did for
everything else (*different shape*).

Ask yourself these two questions:

1. What is the responsibility of the quantity method?
2. Is there a way to make the fourth phrases more alike, even if not yet
   identical?

When you’re confused, don’t try to solve the entire problem
straightaway. The more confused you are, the more important it is to
nibble.

Reminder of Flocking Rules:

1. Select the things that are most alike.
2. Find the smallest difference between them.
3. Make the simplest change that will remove that difference.

After step 3 of Flocking: as always, you
must name this concept, create a method, and send the message in place
of the difference.

The above is the iterative application of Flocking Rules.

> Successor is important, and separating it from quantity gives both
> methods a single responsibility. If you conflate
> choosing-what-to-sing-for-any-number (quantity) with deciding-
> what-verse-to-sing-next (successor), the resulting method would be
> harder to understand, future refactorings would be more difficult, and
> attempts to change the code for one idea might accidentally break it
> for the other.